Remove "Requirements to cover in Issues" section
I feel that this section is redundant. After creating the Epic, the author should immediately create the issues and add them to the Epic. The issue titles will then be listed on the Epic immediately below the main Epic description. If the "Requirements to cover in Issues" section is just a repeat of the issue titles, which IMHO it pretty much always is, then creating it is a waste of time.